Add files via upload
This personal portfolio website is meticulously crafted using HTML, CSS, and JavaScript, showcasing a professional and visually appealing design. It features a responsive layout that adapts seamlessly to various devices, ensuring an optimal user experience. The use of modern web technologies allows for interactive elements and smooth navigation, highlighting the individual's skills, projects, and achievements effectively. This portfolio serves as a powerful tool for personal branding, enabling users to present their work in a compelling manner.